BF: Base scanpath simplification on per-iteration changes
Previously, the amount of iterations during simplification was based on the length of the scanpath.
Scanpath simplification stopped if the number of iterations was equal to the scanpath length.
This logic fails in cases where the scanpath length drops lower than the amount of iterations without ever equaling it (#46)
This change bases the decision to terminate the simplification on the condition that no length reduction
(and thus no simplification) has happened in the previous iteration.

Thanks to @LFaggi for the bug report and code patch. Fixes #46